updated xcode project, fixed some warnings and errors

+++ src/moaiext-audiosampler/MOAIAudioSampler.cpp
@@ -148,7 +148,7 @@ int	MOAIAudioSampler::_prepareBuffer ( lua_State* L ) {
         return 0;
     }
 
-    int bufsize = MyComputeRecordBufferSize( &self->recFmt, self->queue, sec );
+    int bufsize = MyComputeRecordBufferSize( &self->recFmt, self->queue, ( float )sec );
     if( bufsize < 0 ){
         printf("invalid arg?");
         return 0;
